+++ src/index.html
@@ -41,7 +41,7 @@
 </div>
 
 <div class="popup-edit popup">
-  <form action="#" class="form popup__form" name="form_edit" novalidate>
+  <form action="#" class="form popup__form" name="form__edit" novalidate>
     <h2 class="form__title">Редактировать профиль</h2>
     <label class="form__label">
       <input type="text" id="input-name" class="form__input form__input_edit_name" name="edit_name_avatar"
